filled out model class methods

+++ Swift Roster/Person.swift
@@ -66,9 +66,9 @@ class Person: NSObject {
     }
     
     func randomColor() -> UIColor {
-        let r = random().bridgeToObjectiveC().floatValue
-        let g = random().bridgeToObjectiveC().floatValue
-        let b = random().bridgeToObjectiveC().floatValue
+        let r = random().bridgeToObjectiveC().doubleValue
+        let g = random().bridgeToObjectiveC().doubleValue
+        let b = random().bridgeToObjectiveC().doubleValue
         return UIColor(red: r, green: g, blue: b, alpha: 1.0)
     }
 }
